completed point value integration

+++ core/app/Http/Utils/CartService.php
@@ -12,7 +12,13 @@ namespace App\Http\Utils;
 use App\Cart;
 use App\CartItem;
 use App\Product;
+use App\User;
 
+const UNORDERED = 0;
+const PENDING = 1;
+const COMPLETED = 2;
+const FAILED = 3;
+const REVIVE = 4;
 class CartService
 {
     public $productModel;
@@ -29,7 +35,13 @@ class CartService
         if($current_cart) {
             return $current_cart;
         }
-        $cart = new Cart();
+        $reference_no = getTrx();
+        while (Cart::where('reference_no', $reference_no)->get()->count()) {
+            $reference_no = getTrx();
+        }
+        $cart = new Cart([
+            'reference_no' => $reference_no
+        ]);
         $customer->carts()->save($cart);
         return $cart;
     }
@@ -61,24 +73,27 @@ class CartService
         $cartItem = new CartItem([
             'quantity' => $qty,
             'product_id' => $product->id,
+            'point_value' => $product->point_value,
             'price' => $product->promo_price ?? $product->price,
             'weight' => $product->weight ?? 0
         ]);
         $cI = $cart->cartItems()->save($cartItem);
         return !!$cI;
     }
-    public function getCartMetaData(Cart $cart) {
-        $cartTotal=0;
+    public function getCartMetaData(Cart $cart): CartMetaData {
+        $cartTotal=$cart->shipping;
+        $pointValue=0;
         $weightTotal=0;
         $map = [];
         $mapKeys = [];
-        $cart->cartItems()->get()->each(function ($item) use(&$map, &$mapKeys, &$cartTotal, &$weightTotal) {
+        $cart->cartItems()->get()->each(function ($item) use(&$map, &$mapKeys, &$cartTotal, &$weightTotal, &$pointValue) {
             $cartTotal+= ceil($item->quantity * $item->price);
+            $pointValue+=($item->point_value * $item->quantity);
             $weightTotal+=$item->weight;
             $map[$item->product_id] = $item->quantity;
             $mapKeys[] = $item->product_id;
         });
-        return [$mapKeys, $map, $cartTotal, $weightTotal];
+        return new CartMetaData($mapKeys, $map, $cartTotal, $weightTotal, $pointValue);
     }
     public function syncProductsWithCartQuantity($products, $map, $mapKeys) {
         $products->map(function ($item) use ($mapKeys, $map) {
@@ -92,91 +107,115 @@ class CartService
         return $products;
     }
     public function getProductsAndCartTotal($products) {
-        [$mapKeys, $map, $cartTotal] = $this->getCartMetaData( $this->cart());
-        $products = $this->syncProductsWithCartQuantity($products, $map, $mapKeys);
-        return [$products, $cartTotal];
+        $cartMetaData = $this->getCartMetaData( $this->cart());
+        $products = $this->syncProductsWithCartQuantity($products, $cartMetaData->map, $cartMetaData->mapKeys);
+        return [$products, $cartMetaData->cartTotal, $cartMetaData->pointValue];
     }
     public function checkout($address, $other_info) {
         $cart = $this->cart();
         $user = auth()->user();
         if (!$user->my_level()->first()) {
+            $notify[] = ['error', 'You need to purchase a plan to start buying products'];
             return redirect()->route('user.plan.purchase')->withNotify($notify);
         }
-        [, , $cartTotal] = $this->getCartMetaData($cart);
-        if($cartTotal > $user->balance) {
+        $cartMetaData = $this->getCartMetaData($cart);
+        if($cartMetaData->cartTotal > $user->balance) {
             $notify[] = ['error', 'Insufficient balance'];
             return back()->withErrors($notify);
         }
-        $user->update(['balance' => $user->balance - $cartTotal]);
-        $data = ['status' => 1, 'address' => $address, 'other_info' => $other_info];
+        $user->update(['balance' => $user->balance - $cartMetaData->cartTotal]);
+        $this->logTransaction($user, $cartMetaData->cartTotal, 'Fresh Order payment', 'payment', $cart->reference_no ?? 'FS677D79SHS');
+        $data = ['status' => PENDING, 'address' => $address, 'other_info' => $other_info];
         $cart->update($data);
         $cart->cartItems()->each(function (CartItem $cartItem) {
-            $cartItem->update(['status' => 1]);
+            $cartItem->update(['status' => PENDING]);
         });
         $notify[] = ['success','Order initiated successfully'];
         return redirect()->route('user.orders-pending')->withNotify($notify);
 
     }
     public function getUserCarts(int $status) {
-        $statuses = [0 => 'Un ordered', 1 => 'Pending', 2 => 'Completed', 3 => 'Failed'];
+        $statuses = [UNORDERED => 'Un ordered', PENDING => 'Pending', COMPLETED => 'Completed', FAILED => 'Failed'];
         $carts =  auth()->user()->carts()->where('status', $status)->get();
         $carts->each(function (Cart $cart) use ($statuses) {
-            [$mapKeys, $map, $cartTotal, $weightTotal] = $this->getCartMetaData($cart);
-            $cart->weight = $weightTotal;
-            $cart->amount = $cartTotal;
+            $cartMetaData = $this->getCartMetaData($cart);
+            $cart->weight = $cartMetaData->weightTotal;
+            $cart->amount = $cartMetaData->cartTotal;
+            $cart->pointValue = $cartMetaData->pointValue;
             $cart->status = $statuses[$cart->status];
             return $cart;
         });
         return $carts;
     }
     public function getUsersCarts(int $status) {
-        $statuses = [0 => 'Un ordered', 1 => 'Pending', 2 => 'Completed', 3 => 'Failed'];
+        $statuses = [UNORDERED => 'Un ordered', PENDING => 'Pending', COMPLETED => 'Completed', FAILED => 'Failed'];
         $carts =  Cart::where('status', $status)->orderBy('created_at', 'desc')->get();
         $carts->each(function (Cart $cart) use ($statuses) {
-            [$mapKeys, $map, $cartTotal, $weightTotal] = $this->getCartMetaData($cart);
-            $cart->weight = $weightTotal;
-            $cart->amount = $cartTotal;
+            $cartMetaData = $this->getCartMetaData($cart);
+            $cart->weight = $cartMetaData->weightTotal;
+            $cart->amount = $cartMetaData->cartTotal;
+            $cart->pointValue = $cartMetaData->pointValue;
             return $cart;
         });
         return $carts;
     }
     public function updateCartStatus(Cart $cart, int $status) {
+        if($cart->status === $status) {
+            $notify[] = ['success', 'Action already performed'];
+            return back()->withNotify($notify);
+        }
+        if($cart->status === COMPLETED && $status === FAILED) {
+            $notify[] = ['error', 'You can only fail an in complete order'];
+            return back()->withNotify($notify);
+        }
+        if ($cart->status !== FAILED && $status === REVIVE) {
+            $notify[] = ['error', 'Order was n\'t failed. Cannot revive order'];
+            return back()->withNotify($notify);
+        }
         $user = $cart->user()->first();
+        $cartMetaData = $this->getCartMetaData($cart);
         switch ($status) {
-            case 3: // failed
-                $cartTotal = $cart->shipping;
+            case FAILED: // failed
+                $cartTotal = ceil($cart->shipping);
                 $cart->cartItems()->each(function (CartItem $cartItem) use ($status, &$cartTotal) {
-                    $cartItem->status = $status;
-                    $cartTotal+=$cartItem->price;
+                    $cartItem->update(['status' => $status]);
+                    $cartTotal+=($cartItem->price * $cartItem->quantity);
                     return $cartItem;
                 });
-                $user->update(['balance' => $user->balance + ceil($cartTotal)]);
-                $cart->update(['status' => $status]);
-                $notify[] = ['success', 'Order status changed successfully'];
+                if($cart->status === COMPLETED) {
+                    $user->update(['point_value' => $user->point_value - $cartMetaData->pointValue]);
+                }
+                $user->update(['balance' => $user->balance + $cartTotal]);
+                $this->logTransaction($user, $cartTotal, 'Order refund', 'refund', $cart->reference_no ?? 'FS677D79SHS');
+                $cart->update(['status' => FAILED]);
+                $notify[] = ['success', 'Order refund successful'];
                 break;
-            case 4: // revived
-                [, , $cartTotal, ] = $this->getCartMetaData($cart);
-                if ($user->balance < $cartTotal) {
+            case REVIVE: // revived
+                if ($user->balance < $cartMetaData->cartTotal) {
                     $notify[] = ['error', 'User balance cannot revive order'];
                     break;
                 }
-                $cartTotal = $cart->shipping;
+                $user->update(['balance' => $user->balance - $cartMetaData->cartTotal]);
+                $this->logTransaction($user, $cartMetaData->cartTotal, 'Order payment', 'payment', $cart->reference_no ?? 'FS677D79SHS');
                 $cart->cartItems()->each(function (CartItem $cartItem) use ($status, &$cartTotal) {
-                    $cartItem->status = 1;
-                    $cartTotal+=$cartItem->price;
+                    $cartItem->update(['status' => PENDING]);
                     return $cartItem;
                 });
-                $user->update(['balance' => $user->balance - ceil($cartTotal)]);
-                $cart->update(['status' => 1]);
-                $notify[] = ['success', 'Order revived successfully'];
+                $cart->update(['status' => PENDING]);
+                $notify[] = ['success', 'Order revived successfully check pending orders'];
                 break;
             default:
-                if($cart->status === 3) {
-                    $notify[] = ['error', 'User has not paid pleas revive order and try again!'];
+                if($cart->status === FAILED) {
+                    $notify[] = ['error', 'User has not paid please revive order and try again!'];
                     break;
                 }
+                if($status === COMPLETED && in_array($cart->status, [PENDING]) ) {
+                    $user->update(['point_value' => $user->point_value + $cartMetaData->pointValue]);
+                } else if(($cart->status === COMPLETED) && ($status  && !in_array($status, [COMPLETED, REVIVE, FAILED]))) {
+                    $user->update(['point_value' => $user->point_value - $cartMetaData->pointValue]);
+                }
                 $cart->cartItems()->each(function (CartItem $cartItem) use ($status) {
-                    $cartItem->status = $status;
+                    $cartItem->update(['status' => $status]);
                     return $cartItem;
                 });
                 $cart->update(['status' => $status]);
@@ -191,13 +230,26 @@ class CartService
     public function updateCartItemStatus(Cart $cart, CartItem $cartItem, int $status) {
         $user = $cart->user()->first();
         switch ($status) {
-            case 3: // failed
-                $charges = $cartItem->price;
+            case FAILED: // failed
+                if($cart->status === FAILED) {
+                    $notify[] = ['error', 'User has not paid please revive order and try again!'];
+                    break;
+                }
+                $charges = ceil($cartItem->price);
+                if($cartItem->status === COMPLETED) {
+                    $pointValue = $cartItem->point_value * $cartItem->quantity;
+                    $user->update(['point_value' => $user->point_value - $pointValue]);
+                }
                 $cartItem->delete();
-                $user->update(['balance' => $user->balance + ceil($charges)]);
+                $user->update(['balance' => $user->balance + $charges]);
+                $this->logTransaction($user, $charges, 'Order item refund for '.$cartItem->product->name, 'refund', $cart->reference_no ?? 'FS677D79SHS');
                 $notify[] = ['success', 'Order item removed successfully'];
                 break;
             default:
+                if($cart->status === FAILED) {
+                    $notify[] = ['error', 'User has not paid please revive order and try again!'];
+                    break;
+                }
                 $cartItem->update(['status' => $status]);
                 $notify[] = ['success', 'Order item status changed successfully'];
                 break;
@@ -207,5 +259,35 @@ class CartService
         }
         return back()->withNotify($notify);
     }
+    public function logTransaction(User $user, $amount, $title, $type, $transaction_no) {
+        $user->transactions()->create([
+            'amount' => $amount,
+            'user_id' => $user->id,
+            'main_amo' => $amount,
+            'balance' => $user->balance,
+            'type' => $type,
+            'title' => $title,
+            'trx' => substr(getTrx(), 0,6).substr($transaction_no, 0,6),
+        ]);
+    }
 
+}
+
+
+
+class CartMetaData {
+    public $mapKeys;
+    public $map;
+    public $cartTotal;
+    public $weightTotal;
+    public $pointValue;
+
+    public function __construct($mapKeys, $map, $cartTotal, $weightTotal, $pointValue)
+    {
+        $this->mapKeys = $mapKeys;
+        $this->map = $map;
+        $this->cartTotal = $cartTotal;
+        $this->weightTotal = $weightTotal;
+        $this->pointValue = $pointValue;
+    }
 }
